+++ src/netspresso_trainer/pipelines/task_processors/detection.py
@@ -18,6 +18,8 @@ from typing import Literal, Optional
 
 import torch
 
+from netspresso_trainer.models.utils import set_training_targets
+
 from .base import BaseTaskProcessor
 
 
@@ -40,12 +42,7 @@ class DetectionProcessor(BaseTaskProcessor):
         optimizer.zero_grad()
 
         with torch.cuda.amp.autocast(enabled=self.mixed_precision):
-            if isinstance(train_model, torch.nn.parallel.DistributedDataParallel):
-                if hasattr(train_model.module.head, "_training_targets"): # for RT-DETR
-                    train_model.module.head.set_training_targets(targets)
-            else:
-                if hasattr(train_model.head, "_training_targets"):
-                    train_model.head.set_training_targets(targets)
+            train_model = set_training_targets(train_model, targets)
             out = train_model(images)
             loss_factory.calc(out, targets, phase='train')