+++ src/netspresso_trainer/pipelines/train.py
@@ -288,7 +288,7 @@ class TrainingPipeline(BasePipeline):
 
         model = copy.deepcopy(model)
 
-        save_dtype = model.save_dtype
+        save_dtype = model.save_dtype if hasattr(model, 'save_dtype') else next(model.parameters()).dtype
         if save_dtype == torch.float16:
             model = model.type(save_dtype)
 
@@ -333,7 +333,7 @@ class TrainingPipeline(BasePipeline):
         model = self.model.module if hasattr(self.model, 'module') else self.model
         best_model = copy.deepcopy(model)
 
-        save_dtype = best_model.save_dtype
+        save_dtype = best_model.save_dtype if hasattr(best_model, 'save_dtype') else next(best_model.parameters()).dtype
         if save_dtype == torch.float16:
             best_model = best_model.type(save_dtype)
 
@@ -355,6 +355,9 @@ class TrainingPipeline(BasePipeline):
                     sample_input=self.sample_input.type(save_dtype),
                     opset_version=opset_version)
             logger.info(f"ONNX model converting and saved at {str(model_save_path.with_suffix('.onnx'))}")
+            if self.logger.use_mlflow:
+                self.logger.mlflow_logger.log_onnx_model(model_save_path.with_suffix('.onnx'),
+                input_example=self.sample_input.type(save_dtype))
 
             if not self.is_graphmodule_training:
                 save_graphmodule(best_model,